Merge the Apache 2.2 and 2.4 configurations
In preparation for a clean-up of our Apache configurations[0], merging
the Apache configurations seems wise. ``mod_version``[1] appears in
Apache 2.0 onward, so we can use it to do if/else blocks in the configs.
There are very few differences between the two configurations so it
makes sense to maintain just one.

[0] https://pulp.plan.io/issues/1257
[1] https://httpd.apache.org/docs/2.4/mod/mod_version.html