_release_resource cleans up killed worker tasks
If a child Celery process is killed, Celery would recover
but the TaskStatus would stay at running forever. This adds
an additional check when the resource lock is removed which
handles cleanup for any worker being killed.

Because the _release_resource is being run, we know the task
itself has exited, yet if the TaskStatus is not at a final state,
the worker certainly errored.

https://pulp.plan.io/issues/1673
fixes #1673