Backport proposed requests HTTPAdapter
This contains the changes to the default HTTP adapter for the requests
library proposed in GitHub pull request #3109. Note that at this time,
that pull request has not been accepted and is subject to change. In
addition to the modified adapter, the Pulp streamer (and only the Pulp
streamer) has been changed to use this adapter.

Note this fix only works if urllib-1.16+ is used in conjunction with
requests.

closes #1788