Re-apply changes that weren't correctly merged forward
A bug in our merge-forward code caused the commits making these
changes to be recorded as applied to master, but the actual
code changes were not made due to the nature of the merge.

This commit restores the code changes that should have taken
place were a "proper" merge-forward done.