Pull basic auth info out of feed url
Previous versions of Pulp allowed users to include their basic auth
information in the url. This is obviously less ideal than passing
this in the config, but in order to remain backwards compatibile, this
removes the auth info from the url and adds it to the importer config.
For importer configurations that were created before this change, we pass
the configuration through the same cleaner function as we do during
importer creation, so a temporary importer configuration includes the
basic auth information.

closes #1977