Fix crash from removing missing option from dist conf
Fixed crash that arise when we try to remove an option from
distributor confing which was not before set. This ends up
in KeyError because this option was not in config dict.
This was fixed and also added TestCase for it.

Summary:
* Fixed pulp.server.controllers.distributor.update to ignore
missing options from distributor config
* Added TestCase for this behaviour
TestUpdate.test_remove_none_values_non_existing()

closes #2134
https://pulp.plan.io/issues/2134