Implement a useful download API for Pulp
This API is intended to replace the Nectar library. It uses
requests-futures, a library that wraps requests with Python 3.2's
`concurrent.futures`. Rather than returning responses, this library
returns futures. It also provides a simply callback to allow users
to provide a function that works with the response in a background
thread.

By default, it uses a ThreadPoolExecutor, but in Python 3.4+ it is
possible to use a ProcessPoolExecutor (with a few caveats).

closes #2168