Explicitly set view_name when registering NamedModelViewSets
Without this, we had an inconsistency where a ViewSet, e.g. "TasksViewSet"
with endpoint_name "tasks" would still be registered with a base_name of
"task", so the section of the docs that say to use endpoint_name when
generating view_name arguments to serializer reference fields is a liar
without this change.