Restore platform API docs
After the work in https://pulp.plan.io/issues/2327, we did a docs
restructure that didn't include that work. This restores and improves
upon it, by including the entire platform API as it stands today.

Minor changes had to be made to some modules to get the docs to build,
including correcting sphinx syntax errors and fixing some imports that
are no longer valid.

re #2327