Updates build steps related to docs
The build docs were missing the setting of the
LATEST variable which caused new GA releases to
not be published at the root of docs.pulpproject.org

While in there, I tried to improve the clarity and ensure
each line is not over 100 characters.

I built the docs locally and proofed them also.