Problem: pulp_celerybeat writes hearbeat records to the database
Solution: All workers write their own records to the db instead of
sending them through the message bus.

This patch makes use of the Consumer blueprint that celery runs at
start time of a worker. An extra bootstep has been added to the
consumer blueprint. The new bootstep sets a 30 second timer which
runs a method that records a heartbeat in the database.

http://docs.celeryproject.org/en/master/userguide/extending.html
https://groups.google.com/d/msg/celery-users/3fs0ocREYqw/C7U1lCAp56sJ

closes #2519
https://pulp.plan.io/issues/2519