Problem: RSyncFastForwardUnitPublishStep publishes incorrectly
Solution: Enable RSyncFastForwardUnitPublishStep to publish the same type of units into different
directories by extracting the publish path of a unit from it's symlink location.

This problem was introduced when the yum distributor was modified to publish repositories with a
new layout. This patch makes the RPM rsync distributor consistent with the yum distributor.

re #1976
https://pulp.plan.io/issues/1976