Update testing certificates for oid_validation and repoauth
It looks worse than it is. m2crypto has started to reject any
certificate with and expiration data past 2050, which was the case with
our testing certificates here. Since I fully expect we'll want to bring
the repoauth and oid validation features forward to Pulp 3, I thought it
was worth scripting up the openssl-fu necessary to easily remake these
certificates in the future.

I reorganized the test_oid_validation file significantly, but the only
real functional change was conflating the ideas of "VALID_CA2" and
"INVALID_CA". There was actually nothing invalid about INVALID_CA; the
only thing invalid about it was that it wasn't VALID_CA. This was also
true for VALID_CA2, so they got merged into "OTHER_CA", since (as far
as I could tell) there was no test that wanted to use both INVALID_CA
and VALID_CA2 at the same time. Most of the churn comes from loading
these certs from the filesystem instead of embedding them as strings in
the test module. The repoauth tests already worked this way, but don't
have as many testing certs, so the regenerate script also does repoauth
a solid and updates its certs while it's freshening everything up.

Finally, I renamed the contants to be a little more consistent within
the test module, and removed constants aren't being used anywhere.