Workers write their own hearbeat records to database.
All workers will write their own records to the database
instead of relying on pulp_celerybeat to do so for them
using celery heartbeats.

This patch makes use of the Consumer blueprint that celery
runs at the start time of a worker. An extra boot step has
been added which sets a timer to periodically update the
worker record in the database.

http://docs.celeryproject.org/en/master/userguide/extending.html
https://groups.google.com/d/msg/celery-users/3fs0ocREYqw/C7U1lCAp56sJ

closes #2519
https://pulp.plan.io/issues/2519
closes #2516
https://pulp.plan.io/issues/2516