Problem: Content can not be created using REST API
Solution: Add viewset and serializers to serve Content creation API

This patch also introduces a DeferredArtifact that will be used by importers that
want to support deferred download policies. This model replaces the DownloadCatalog
model. However, DownloadCatalog was not removed as part of this patch because it is
referenced by the changesets. The changesets will be updated in a separate commit and
the DownloadCatalog will be removed then.

This patch also includes a fix for the artifact_pre_delete signal which should have
been updated when the Artifact model changed.

This patch also introduces additional base viewsets for models that don't need update()
implemented. REST API for both Artifact and Content resources does not support updates.

closes #2872
https://pulp.plan.io/issues/2872