Add Distribution ViewSet and Serializer
Additional work was done to allow arbitrarily deeply nested ViewSets.

The router registration code in urls.py was significantly changed.
A `ViewSetNode` tree is assembled that mirrors the nesting hierarchy
of the url structure, and is able to create child routers as necessary
and use them to register child ViewSets.

A `parent_viewset` attribute is added to ViewSets, which indicates
how they are nested and assists in the retrieval of parents.

A new type of field, `WritableNestedUrlRelatedField` was added to
allow parent fields to be writable when the parent is determined by
the nested url rather than by request parameters (and designated
`read_only`). Serializers with writable url fields inherit from
`NestedModelSerializer`.