The FileSystemStorage backend provided by Django has been extended to conform to Pulp needs. When saving a file, the storage backend does nothing if a file by that name already exists. All TemporaryDownloadedFile and TemporaryUploadedFile objects are moved using os.rename() when possible.
The new ArtifactFileField allows plugin writers to instantiate an Artifact with a string path to a downloaded
file. This helps the plugin writer avoid having to open and close a downloaded file. The pre_save()[0] method
instantiates a TemporaryDownloadedFile object and passes it to the storage backend to save into place. The
save()[1] method on the Artifact then closes the file opened by pre_save().

The get_artifact_path() and published_metadata_path() static methods are now helper methods that can be used
with any storage backend. In particular, I expect these to work with S3 storage backend without any
modification.

The signal handler for pre_delete of the Artifact model has also been removed. This handler relied on the
custom storage backend to remove empty directories left behind after an Artifact has been removed. This is
not necesary because there can only be 255 directories inside /var/lib/pulp/artifact. It is highly unlikely that a
directory will remain empty for long.

[0] https://docs.djangoproject.com/en/1.11/ref/models/fields/#django.db.models.Field.pre_save
[1] https://docs.djangoproject.com/en/1.11/ref/models/instances/#django.db.models.Model.save

closes: #2950
https://pulp.plan.io/issues/2950