Fix the api schema route
The issue was that we were mapping all /api/v3/ routes to the api schema
which meant that all routes that didn't match another route displayed
the api schema. It also caused some weird behaviors where we saw some
406s instead of 404. I also needed to move the route above the other
routers because they assigned /api/v3/ to a route which returns a list
of resources.

fixes #3063
https://pulp.plan.io/issues/3063