Problem: Distribution lacks a base URL
Solution: Add 'base_url' field to the Distribution resource

This patch adds a read only 'base_url' field to the Distribution resource. It's value is derived
from a new 'host' config in server.yaml and the 'base_path' of the Distribution. If the 'host' is
not set, the host is is assumed to be the same the one in the request.

closes #3185
https://pulp.plan.io/issues/3185