Fixes workers crashing on restart edge case
Fixes an issue where workers would crash when
attempting to revoke() a previously running task
on startup. If the user sends SIGKILL to the
worker while running a task, the worker will
attempt to mark the task as stale when restarted.
This causes a hang for yet-unknown reasons. Calling
revoke() is not necessary when a dead worker is being
cleaned up because all work goes through dedicated
queues which disappear when the worker does. We
leave revoke() active only when a task is being
cancelled outside of worker cleanup.

closes #2835
https://pulp.plan.io/issues/2835

	modified:   server/pulp/server/async/tasks.py
	modified:   server/test/unit/server/async/test_tasks.py

(cherry picked from commit aa7a1de219b02bb0aa5a3674a72d77a653ee968f)