Have get_resource() use pk and only pk if pk is provided
We should ignore other kwargs since pk should uniquely identify the
resource. The problem we were experiencing in pulp_ansible is that we
had a pk and a role_pk. The conversion from role_pk to role_id only
happens inside the else block and so we're getting a FieldError because
role_pk doesn't exist on the model. Instead, I think we should just use
the pk kwarg and ignore the rest.