Adds static content docs
This adds static content configuration docs, and adds it as a component
in the architecture list. It adds a single step to the installer which
refers to the more detailes docs.

It also adds the 'DEBUG: False' as the default to settings.yaml which
was missing. It was needed because these docs mention configuring it
there.

It also makes a few formatting improvements.

https://pulp.plan.io/issues/3620
closes #3620