Consistent names in server.yaml
The naming convention (not required) for yaml is lowercase_snake.
We had some ALL_UPPERCASE names mixed in and wanted consistency.
Due to some Django expectations, nested dicts, etc, we chose to
make the names in server.yaml ALL_UPPERCASE with the exception
of the logging section as Python is expecting lowercase there.

fixes #3589
https://pulp.plan.io/issues/3589