Fix task in resource manager queue not cancel.
When stopping resource manager worker, tasks that assigned to the
resource manager will not be cancelled and leaving the tasks in
"Waiting" state forever. This is because the "worker_name" of
the TaskStatus is not set during the creation and Pulp is unable
to find the tasks for the resource manager during the worker
cleanup.