Remove usages of selectors.bug_is_fixed()
Now that Smash tests are in the same repo as the code under test, we no
longer need to look to an external source (Redmine) to determine whether
an issue should be tested.

When adding a new test for funtionality which does not exist or is broken, decorate the test or test class with @unittest.skip("https://pulp.plan.io/issue/..."), and the patch which adds or fixes this functionality should remove the skip decorator to re-enable the test.

This has several benefits - it means the PR test runner will always run
the tests for the functionality being added/fixed so long as the
developer has properly re-enabled the test. It should also make the
tests faster by a small degree, and less reliant on external connections.