Fixes broken unit test around TaskStatus model
The test uses a list of various objects as invalid inputs for a Task status. One of the objects is a dict
which mongoengine does not support as a value because it's not hashable. This patch removes the dict from
the list of possible bad inputs.

closes: #3955
https://pulp.plan.io/issues/3955
(cherry picked from commit fe93f013deba078c7d8e927dfef0476f1216c33f)