Problem: OpenAPI schema for RepositoryVersion 'retrieve' is wrong
Solution: Create separate serializers for 'create' operation

The problem was that we had a single serializer for both GET and POST. DRF was able to
distinguish which fields were to be used for accepting POST body params and which fields to
use when serializing for a GET response. DRF knows this because some fields are 'write_only'.
However, the drf_yasg library we use for generating OpenAPI schema does not take into account
the 'write_only' attribute of a field. As a result when it was using the serializer to
generate a schema for the GET response it was including the 'write_only' fields.

This patch introduces a new serializer that is only used for the 'create' operation of the
RepositoryVersion ViewSet. This serializer is only used to inspect the inbound data in the POST
body. It is not used for the response. The response is generated using the OperationPostponed
serializer.

closes: #3966
https://pulp.plan.io/issues/3966