Problem: mirror is defined as a string in OpenAPI schema
Solution: change the ChoiceField to a BooleanField

Auto generated bindings were sending a string 'true' for the default. The REST API would respond
with a 400 error. Since the two choices were 'true' and 'false', it was more appropriate to use a
boolean field. The OpenAPI schema now correctly describes this field as a boolean.

[noissue]