Problem: href values in browsable API are not clickable
Solution: add a filter that makes links out of href values

This patch adds a copy of the rest_framework api template to this repository. This was necesary in order to use
the new custom "urlize_quoted_links" filter. The filter turns any string that starts with '/pulp/api/v3' into a
anchor tag with that value.

closes #3987
https://pulp.plan.io/issues/3987