Problem: ArtifactFileField can delete an existing artifact's file
Solution: Add a check to see if the source file is already in artifact storage

Prior to this patch it was possible to break an existing artifact by passing a path to it's file
as the file for a new Artifact.

This patch also removes extraneous code in the ArtifactSaver stage. The code was present because
an assumption was made that bulk_save() on Artifacts would not call the presave handler on the
ArtifactFileField. However, this is not the case and the pre_save handler is getting called.

closes: #3915
https://pulp.plan.io/issues/3915