Problem: database settings are not being used
Solution: load dynaconf in manage.py, wsgi.py, worker.py, and entry_points.py

This patch introduces a workaround for a dynaconf issue[0].

This patch also changes how Pulp is deployed on Travis. A non-default name is used for the database. That
way we will know if the settings in the settings file are being used.

[0] https://github.com/rochacbruno/dynaconf/issues/89

closes: #4046
https://pulp.plan.io/issues/4046