Add tests for creating resources using invalid data
Add a test to verify messages when attempting to create a user with a long
`username` and `username` with invalid characters. Also, verify that HTTP error
happening in both cases.

Besides that, refactor the test for create repos with the same name.

See: https://github.com/PulpQE/pulp-smash/issues/882

'[noissue]'