Close HttpDownloader connections properly
The close() method is a coroutine so it needs to be awaited but it was
not. This was discovered during some testing of the HttpDownloader I was
doing outside of Pulp.

It also fixes some of the examples which used the wrong class name and
an await example that would have errored if run.

[noissue]