Adds 'policy' to Remote
Required PR: https://github.com/pulp/pulp_file/pull/132

This includes a serializer that validates input correctly.

This also switches the finalize, write_data, and handle_headers
callbacks to be coroutines themselves. This is required because within
those callbacks the streamer needs to await on coroutines and that can
only be done from withing a coroutine itself.

This also adjusts the Remote.get_downloader() method signature to take
`url` or `remote_artifact` as kwargs instead of `url` as a positional
parameter. This allows plugin writers (and the streamer) to get
downloaders configured with digest and size validation automatically
from remote_artifacts.

https://pulp.plan.io/issues/3763
closes #3763