Problem: bulk_create can fail with IntegrityError
Solution: add a manager that provides a bulk_get_or_create() method

This patch introduces a BulkCreateManager that provides the bulk_get_or_create method. This method
handles IntegrityErrors encountered during bulk_create() by inserting each object into the database
serially. When an IntegrityError or ValueError is encountered during serial saving of objects, the
object being saved is replaced with an instance from the database.

This patch introduces a mixin used by Artifact, Content, ContentArtifact, and RemoteArtifact models.
The mixin provides the q() method which returns a Q object that can be used to retreive the database
instance of the model.

closes #4060
https://pulp.plan.io/issues/4060