Fixing missing hostname for deferred downloads
A change was made in #4092 to allow users to use the server request
hostname as the redirect_host for on-demand repos by leaving host in the
lazy config as blank. However, the defered_download code doesn't have a
server request and therefore should default to socket.getfqdn() if
redirect_host is unset.

fixes #4120
https://pulp.plan.io/issues/4120