Adds the content app to pulpcore.content
The Settings can be a lot simpler now that the architecture no longer
required redirection. It does need to account for the content being
served on a different HOST. This PR handles that.

Since the settings are now single values, they are much easier to
override using Dynaconf, which is great.

The content app can be removed now that the streamer code is handling it
all.

This PR is a prototype and has no docs. Use Pulp as normal , but run the
streamer from github.com/bmbouter/pulp_streamer/ and run the streamer
with:

`gunicorn pulpcore.streamer:server --bind localhost:8080 --worker-class aiohttp.GunicornWebWorker -w 2`

This PR needs pulp-smash changes from pulp_file so it requires the PR
below:

Required PR: https://github.com/pulp/pulp_file/pull/140

https://pulp.plan.io/issues/4239
closes #4239

https://pulp.plan.io/issues/3698
closes #3698

https://pulp.plan.io/issues/3699
closes #3699

https://pulp.plan.io/issues/4181
closes #4181

https://pulp.plan.io/issues/4243
closes #4243