Do not search for dupes by pk
The ContentUnitSaver stage uses content.q() to retrieve the query for a
duplicate content unit. However, there are cases where the pk of the
content is set even when the save fails with an IntegrityError. This
means that content.q() will always search for content that does not
exist. Instead, we should always query by the natural key.

[noissue]