Fix indention that Shinx complains about
The enqueue_with_reservation is included in the pulpcore-plugin package
and its docstring in the docs. When building the pulpcore-plugin docs it
complains saying:

docstring of pulpcore.plugin.tasking.enqueue_with_reservation
:19:Unexpected indentation.

Also this file has a flak8 issue in it that Travis won't pass until its
fixed.

Also a new version of flake8 causes the comment length to be included
line counts causing unmodified lines to be too long. I shorted them each
slightly.

[noissue]