Adopts new content_summary layout
The performance issue causes us to introduce a new model named
RepositoryVersionContentDetails. This is also a great opportunity to
update the content_summary of the RepositoryVersion serializer to match
the recent API changes.

Most of this code was inspired from a patch from @dalley.

Requires PR: https://github.com/PulpQE/pulp-smash/pull/1174

https://pulp.plan.io/issues/4283
closes #4283