Remove consumer profiles by id
Each profile is a dictionary. It's not reliable to use nested
dictionaries in criteria to remove a document from a collection
because Python doesn't guarantee key order in a dictionary
but it's important for MongoDB.
The solution is to remove by id, the rest of criteria is redundant.

closes #4654
https://pulp.plan.io/issues/4654

(cherry picked from commit dd8c30efd9896f1fa62feb8127e0b10e9a8c6218)