Fixing ISO publish fails with BSON document too large
There's a limit to how large a single mongo query can be (16 Mbytes).
If too many (>134k) units are published, the query to fetch units
will exceed that limit and crash.

This fix sets a threshold to avoid generating too large query and
crash.

If exceeding the threshold, go back to regular full publish since
there won't be much benefit of fast-forward anyway.

ref #5058
https://pulp.plan.io/issues/5058

Limit criteria to fields as needed

Limiting the criteria to fields as needed to save returned result
from the query.