Use multi-resource locking when additional_repos override is passed
The RPM plugin needs to accept more than one source and destination
repository for copy operations. The "additional_repos" flag is being
added to deal with this. Core needs to look for this flag in the
override config and then, when present, lock on all the repo ids present
within as well as the directly-specified source and destination repo
ids.

re: #5067
https://pulp.plan.io/issues/5067