Split filter items to avoid BSON document too large if source repo contains > 345,000 units of same type
The issue was caused by long query (>16M) if source repo contains too many
units ( > 345,000) of same type.

The PR splits unit_ids in the filter to several chunks, then fetches data
separately, eventually combines all results.

Note that the PR only handle the case when the query does not need to sort.
It means the PR will not help if the query contains sorting. So the patch does
not fix the issue pulp#2220 completely.

ref pulp#2220
https://pulp.plan.io/issues/2220