Fix last_publish method call
Method last_publish was not called properly so that a bound
method was used instead of its return. It made publish always
be fast-forward no matter repo has been published or not when
force_full=True is not set.

Added new case for testing this and fixed tests added before.

Closes #5659