Fix $date operator to work with $and or $or
Queries with $date operator doesn't work as expected
with $and or $or operators. $date operator descends
only on dicts and ignores everything else. This fix
modifies $date operator to descend on list too so
it work with $and or $or operators.

Closes #5700

(cherry picked from commit c675c977d922cff3326cc030fe6fffbd7c0274dd)