Adds task failure handling for celery in bad state
Sometimes celery's internal state is in such a bad position it causes
the on_failure handler itself to fail. This is very bad for Pulp as it
causes the tasking system to stop processing tasks due to on_failure not
releasing locks like it should.

This adds a defensive codepath for that situation which maintains the
current functionality, but if Celery is uncooperative it still performs
the on_failure cleanup.

Additionally, it logs the error in all cases. One of the aspects of this
issue is that due to Celery having a bad state it won't correctly log
the actual root cause triggering the `on_failure` handler.

fixes #5502