Set last_publish as a start of publish [RHELDST-5086, #8778]
When a repository is published, set last_publish field on distributor
to the start time of publish instead of finish time. This fix relates
to erraum content type which can be updated even if a repo is in
publish.

With this change pulp shouldn't falsely skip publish of a repo when
last_unit_added is set to time between publish start and finish.

Closes #8778
https://pulp.plan.io/issues/8778